// src/types.ts
export interface Doctor {
  id: string;
  name: string;
  matricula: string;
  matriculaProvincial?: string;
  specialty: string;
  circuloMedico: string;
  avatar?: string;
  email: string;
  phone: string;
  docStatus: 'COMPLETO' | 'PENDIENTE' | 'REVISIÓN';
  placesOfCare: string[];
  conventions: string[];
  documentation: {
    matriculaNacionalFile?: string;
    matriculaNacionalUpdate?: string;
    matriculaNacionalStatus: 'Vigente' | 'No cargada' | 'Vencida';
    matriculaProvincialFile?: string;
    matriculaProvincialStatus: 'Vigente' | 'No cargada' | 'Pendiente';
    seguroMalaPraxisFile?: string;
    seguroMalaPraxisExpiry: string;
    seguroMalaPraxisStatus: 'Vigente' | 'Alerta' | 'Vencido';
  };
  automations: {
    emailAvisoVencimiento: boolean;
    envioResumenContador: boolean;
  };
}
// ...
export type KanbanColumn = 'pendientes' | 'tramite' | 'observadas' | 'aprobadas';

export interface AuthorizationItem {
  id: string;
  code: string;
  patientName: string;
  doctorName: string;
  obraSocial: string;
  timeAgo: string;
  status: KanbanColumn;
  isAutomation: boolean;
  observation?: string;
  progress?: number;
}